completed logout

+++ CRUD-Dapper/Controllers/BranchController.cs
@@ -7,27 +7,47 @@ namespace CRUD_Dapper.Controllers
     public class BranchController : Controller
     {
         private readonly IGenericRepository<Branch> _iGenericRepository;
+
         public BranchController(IGenericRepository<Branch> iGenericRepository)
         {
             _iGenericRepository = iGenericRepository;
         }
 
-        public IGenericRepository<Branch> IGenericRepositroy { get; }
+        // Middleware to check session before action execution
+        private bool IsUserLoggedIn()
+        {
+            return HttpContext.Session.GetString("UserId") != null;
+        }
 
         public async Task<IActionResult> Index()
         {
+            if (!IsUserLoggedIn())
+            {
+                return RedirectToAction("Login", "Auth");
+            }
+
             var result = await _iGenericRepository.GetAll("Branch");
             return View(result);
         }
 
         public async Task<IActionResult> Details(int id)
         {
+            if (!IsUserLoggedIn())
+            {
+                return RedirectToAction("Login", "Auth");
+            }
+
             var result = await _iGenericRepository.GetById("Branch", id);
             return View(result);
         }
 
         public IActionResult Create()
         {
+            if (!IsUserLoggedIn())
+            {
+                return RedirectToAction("Login", "Auth");
+            }
+
             return View();
         }
 
@@ -35,18 +55,27 @@ namespace CRUD_Dapper.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(Branch branch)
         {
+            if (!IsUserLoggedIn())
+            {
+                return RedirectToAction("Login", "Auth");
+            }
+
             if (ModelState.IsValid)
             {
                 branch.CreatedDate = DateTime.Now;
-               
                 await _iGenericRepository.Add("Branch", branch);
-
                 return RedirectToAction(nameof(Index));
             }
             return View(branch);
         }
+
         public async Task<IActionResult> Edit(int id)
         {
+            if (!IsUserLoggedIn())
+            {
+                return RedirectToAction("Login", "Auth");
+            }
+
             var _Branch = await _iGenericRepository.GetById("Branch", id);
             return View(_Branch);
         }
@@ -55,12 +84,16 @@ namespace CRUD_Dapper.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, Branch branch)
         {
+            if (!IsUserLoggedIn())
+            {
+                return RedirectToAction("Login", "Auth");
+            }
+
             if (ModelState.IsValid)
             {
                 try
                 {
                     branch.id = id;
-                   
                     await _iGenericRepository.Update("Branch", branch);
                 }
                 catch (Exception)
@@ -71,8 +104,14 @@ namespace CRUD_Dapper.Controllers
             }
             return View(branch);
         }
+
         public async Task<IActionResult> Delete(int id)
         {
+            if (!IsUserLoggedIn())
+            {
+                return RedirectToAction("Login", "Auth");
+            }
+
             var _Branch = await _iGenericRepository.GetById("Branch", id);
             return View(_Branch);
         }
@@ -80,6 +119,11 @@ namespace CRUD_Dapper.Controllers
         [HttpDelete, ActionName("DeleteConfirmed")]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!IsUserLoggedIn())
+            {
+                return RedirectToAction("Login", "Auth");
+            }
+
             await _iGenericRepository.Delete("Branch", id);
             return RedirectToAction(nameof(Index));
         }
